Move B1517A module to its own module KeysightB1517A.py
...also move parse_spot_measurement_response method into B1517A class
from B1500 class, because that's the only place where this method is
used.
